# target/classes/openapi.yaml
openapi: 3.0.3
info:
  contact:
    email: yermakov80@gmail.com
  description: "Test task: departments and employees"
  title: DepartmentsAndEmployees
  version: 0.0.1
servers:
- description: local server
  url: http://localhost:8080/
...
paths:
  /employees:
    post:
      operationId: addEmployee
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmployeeDTO'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeDTO'
          description: "Successful operation: new employee added"
      summary: Add a new employee
      tags:
      - employees
      x-content-type: application/json
      x-accepts: application/json
      x-tags:
      - tag: employees
    put:
      operationId: editEmployee
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmployeeDTO'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeDTO'
          description: "Successful operation: employee edited"
        "404":
          description: Employee id not found
      summary: Edit employee by employeeID in DTO
      tags:
      - employees
      x-content-type: application/json
      x-accepts: application/json
      x-tags:
      - tag: employees
  /employees/{employeeId}:
    delete:
      operationId: deleteEmployee
      parameters:
      - description: ID of employee to delete
        explode: false
        in: path
        name: employeeId
        required: true
        schema:
          type: integer
        style: simple
      responses:
        "200":
          description: "Successful operation: employee deleted"
        "404":
          description: Employee id not found
      summary: Delete employee by employeeID
      tags:
      - employees
      x-accepts: application/json
      x-tags:
      - tag: employees
    get:
      operationId: getEmployee
      parameters:
      - description: ID of employee to get
        explode: false
        in: path
        name: employeeId
        required: true
        schema:
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeDTO'
          description: "Successful operation: employee data received"
        "404":
          description: Employee id not found
      summary: Get employee by employeeID
      tags:
      - employees
      x-accepts: application/json
      x-tags:
      - tag: employees
  /employees/{employeeId}/{newDepartmentId}:
    patch:
      operationId: editDepartmentOfEmployee
      parameters:
      - description: ID of employee to edit department
        explode: false
        in: path
        name: employeeId
        required: true
        schema:
          type: integer
        style: simple
      - description: ID of department to set
        explode: false
        in: path
        name: newDepartmentId
        required: true
        schema:
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeDTO'
          description: "Successful operation: department of employee edited"
        "404":
          description: Employee id not found
      summary: Set new department id for given employee
      tags:
      - employees
      x-accepts: application/json
      x-tags:
      - tag: employees
...
components:
  schemas:
    EmployeeDTO:
      example:
        employment_date: 2020-03-30T00:00:00.000+00:00
        birth_date: 2000-01-30T00:00:00.000+00:00
        name: Mask Ilon
        id: 12
        id_of_department: 3
        job_title: Manager
      properties:
        id:
          example: 12
          type: integer
        name:
          example: Mask Ilon
          type: string
        birth_date:
          example: 2000-01-30
          format: date
          type: string
        employment_date:
          example: 2020-03-30
          format: date
          type: string
        id_of_department:
          example: 3
          type: integer
        job_title:
          example: Manager
          type: string
      required:
      - name
      type: object
...